Update README.md
Welcome to my personal resume website! This site is a showcase of my skills, experience, education, and projects, presented in an engaging and easy-to-navigate format. Using HTML, CSS, and JavaScript, I've created a fully responsive design that adapts seamlessly to desktop, tablet, and mobile screens. Every piece of content on this site is accurate and up-to-date, ensuring you get a comprehensive view of my professional journey. What sets this website apart is the custom CSS styling that I've crafted, which reflects my unique personality and creativity. I've also added interactivity through JavaScript, incorporating features like a responsive navbar, image sliders, modals, and form validations. You can explore my work and contact me through this platform. I've deployed this website on GitHub Pages to make it easily accessible to everyone, and it functions flawlessly on Chrome, Firefox, Safari, and Edge. Feel free to browse and get to know me better – this website is a testament to my HTML/CSS/JS skills and the key to unlocking my potential as a valuable addition to your team."